moved validator functions to validator class
